+++ zoo-modules/form-modules/input-module/input.spec.mjs
@@ -1,19 +1,4 @@
 describe('Zoo input', function () {
-	it('should create input', async () => {
-		const inputAttrs = await page.evaluate(() => {
-			let input = document.createElement('zoo-input');
-			document.body.appendChild(input);
-			return {
-				inputLabelPresent: input.shadowRoot.querySelector('zoo-label') !== undefined,
-				inputLinkPresent: input.shadowRoot.querySelector('a') !== undefined,
-				inputInfoPresent: input.shadowRoot.querySelector('zoo-info') !== undefined
-			};
-		});
-		expect(inputAttrs.inputLabelPresent).toBeTrue();
-		expect(inputAttrs.inputLinkPresent).toBeTrue();
-		expect(inputAttrs.inputInfoPresent).toBeTrue();
-	});
-
 	it('should pass attributes to input label component', async () => {
 		const labelText = await page.evaluate(() => {
 			document.body.innerHTML = `
@@ -23,7 +8,7 @@ describe('Zoo input', function () {
 					</zoo-input>
 				`;
 			let input = document.querySelector('zoo-input');
-			const label = input.shadowRoot.querySelector('slot[name="label"]').assignedNodes()[0];
+			const label = input.shadowRoot.querySelector('slot[name="label"]').assignedElements()[0];
 			return label.innerHTML;
 		});
 		expect(labelText).toEqual('label');
@@ -35,9 +20,7 @@ describe('Zoo input', function () {
 				<zoo-input>
 					<input id="input-type-number" slot="input" placeholder="input" list="animals"/>
 					<label for="input-type-number" slot="label">Autocomplete</label>
-					<zoo-info slot="info">
-						Possible values: Dog, Cat, Small Pet, Bird, Aquatic
-					</zoo-info>
+					<span slot="info">Possible values: Dog, Cat, Small Pet, Bird, Aquatic</span>
 					<a slot="link" href="https://developer.mozilla.org/en-US/docs/Web/HTML/Element/datalist" target="about:blank">Learn your HTML and don't overcomplicate</a>
 				</zoo-input>
 				`;
@@ -54,28 +37,35 @@ describe('Zoo input', function () {
 		expect(linkAttrs.linkTarget).toEqual('about:blank');
 	});
 
-	it('should accept 1 slot', async () => {
-		const ret = await page.evaluate(() => {
+	it('should render input error', async () => {
+		const errorDisplay = await page.evaluate(() => {
+			document.body.innerHTML = `
+				<zoo-input invalid>
+					<input id="input-type-number" slot="input" placeholder="input" list="animals"/>
+					<label for="input-type-number" slot="label">Autocomplete</label>
+					<span slot="error">error</span>
+				</zoo-input>
+				`;
+			let input = document.querySelector('zoo-input');
+			const error = input.shadowRoot.querySelector('.error');
+			return window.getComputedStyle(error).display;
+		});
+		expect(errorDisplay).toEqual('flex');
+	});
+
+	it('should not render input error', async () => {
+		const errorDisplay = await page.evaluate(() => {
 			document.body.innerHTML = `
 				<zoo-input>
 					<input id="input-type-number" slot="input" placeholder="input" list="animals"/>
 					<label for="input-type-number" slot="label">Autocomplete</label>
-					<zoo-info slot="info">
-						Possible values: Dog, Cat, Small Pet, Bird, Aquatic
-					</zoo-info>
-					<zoo-link slot="link">
-						<a slot="anchor" href="https://developer.mozilla.org/en-US/docs/Web/HTML/Element/datalist" target="about:blank">Learn your HTML and don't overcomplicate</a>
-					</zoo-link>
+					<span slot="error">error</span>
 				</zoo-input>
 				`;
 			let input = document.querySelector('zoo-input');
-
-			const slottedInput = input.shadowRoot.querySelector('slot[name="input"]').assignedNodes()[0];
-
-			return {
-				tagName: slottedInput.tagName
-			};
+			const error = input.shadowRoot.querySelector('.error');
+			return window.getComputedStyle(error).display;
 		});
-		expect(ret.tagName).toEqual('INPUT');
+		expect(errorDisplay).toEqual('none');
 	});
 });