+++ zoo-modules/misc-modules/paginator-module/paginator.js
@@ -20,7 +20,7 @@ export default class Paginator extends HTMLElement {
 	}
 	goToPage(pageNumber) {
 		this.setAttribute('currentpage', pageNumber);
-		this.shadowRoot.host.dispatchEvent(new CustomEvent('pageChange', {
+		this.dispatchEvent(new CustomEvent('pageChange', {
 			detail: {pageNumber: pageNumber}, bubbles: true, compose: true
 		}));
 	}